+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -35,6 +35,7 @@ public class StandardRuleSetProvider : RuleSetProvider {
         NoWildcardImportsRule(),
         PackageNameRule(),
         ParameterListWrappingRule(),
+        SpacingAroundAngleBracketsRule(),
         SpacingAroundColonRule(),
         SpacingAroundCommaRule(),
         SpacingAroundCurlyRule(),