+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The rules below are promoted from the Experiment ruleset to the Standard ruleset
 * `enum-entry-name-case`
 * `package-name`
 * `traling-comma`
+* `spacing-around-angle-brackets`
+* `double-colon-spacing`
 
 Note that as a result of moving the rules that the prefix `experimental:` has to be removed from all references to this rule. Check references in:
 * The `.editorconfig` setting `disabled_rules`.